finishing touches

+++ src/AgendaService.cpp
@@ -2,42 +2,57 @@
 #include <set>
 
 AgendaService::AgendaService() {
-    this->startAgenda();
+    startAgenda();
 }
 
 AgendaService::~AgendaService() {
-    this->quitAgenda();
+    quitAgenda();
 }
 
 bool AgendaService::userLogIn(const std::string &userName, const std::string &password) {
-    return (!m_storage->queryUser([userName, password](const User & x) -> bool {return (x.getName() == userName && x.getPassword() == password);}).empty());
+    return (!m_storage->queryUser([userName, password](const User & x) -> bool {
+        return (x.getName() == userName && x.getPassword() == password);
+    }).empty());
 }
 
 bool AgendaService::userRegister(const std::string &userName, const std::string &password, const std::string &email, const std::string &phone) {
-    if (!m_storage->queryUser([userName](const User &x) -> bool {return (x.getName() == userName);}).empty())
-        return false;  // creating existing user name.
+    if (!m_storage->queryUser([userName](const User &x) -> bool {
+        return (x.getName() == userName);
+    }).empty()) return false;  // creating existing user name.
     m_storage->createUser(User(userName, password, email, phone));
     return true;
 }
 
 bool AgendaService::deleteUser(const std::string &userName, const std::string &password) {
-    if (m_storage->queryUser([userName, password](const User &x) -> bool {return (x.getName() == userName && x.getPassword() == password);}).empty())
-        return false;  // incorrect username or password
+    if (m_storage->queryUser([userName, password](const User &x) -> bool {
+        return (x.getName() == userName && x.getPassword() == password);
+    }).empty()) return false;  // incorrect username or password
+    
     deleteAllMeetings(userName);
     for (auto meeting: listAllParticipateMeetings(userName))
         quitMeeting(userName, meeting.getTitle());
-    m_storage->deleteUser([userName](const User &x) -> bool {return x.getName() == userName;});
+    
+    m_storage->deleteUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    });
+    m_storage->deleteMeeting([](const Meeting &x) -> bool {
+        return x.getParticipator().empty();
+    });
     return true;
 }
 
 std::list<User> AgendaService::listAllUsers(void) const {
-    return m_storage->queryUser([](const User & x) -> bool {return true;});
+    return m_storage->queryUser([](const User & x) -> bool {
+        return true;
+    });
 }
 
 bool AgendaService::createMeeting(const std::string &userName, const std::string &title, const std::string &startDate, const std::string &endDate, const std::vector<std::string> &participator) {
     // user basic check
-    if (m_storage->queryUser([userName](const User &x) {return x.getName() == userName;}).empty())
-        return false;  // the sponsor does not exist
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return false;  // the sponsor does not exist
+    
     if (participator.empty())
         return false;  // no participators
     std::set<std::string> partpool;
@@ -46,8 +61,9 @@ bool AgendaService::createMeeting(const std::string &userName, const std::string
             return false;  // duplicate participator
         if (n == userName)
             return false;  // sponsor is participator
-        if (m_storage->queryUser([n](const User & x) {return x.getName() == n;}).empty())
-            return false;  // participator does not exist
+        if (m_storage->queryUser([n](const User & x) -> bool {
+            return x.getName() == n;
+        }).empty()) return false;  // participator does not exist
         partpool.insert(n);
     }
     
@@ -57,15 +73,18 @@ bool AgendaService::createMeeting(const std::string &userName, const std::string
         return false;  // dates are invalid, or the start date is not earlier than end date
     
     // meeting info basic check
-    if (!m_storage->queryMeeting([title](const Meeting & x) {return x.getTitle() == title;}).empty())
-        return false;  // duplicate title
+    if (!m_storage->queryMeeting([title](const Meeting & x) -> bool {
+        return x.getTitle() == title;
+    }).empty()) return false;  // duplicate title
     
     // free time check
-    if (!m_storage->queryMeeting([userName, sd, ed](const Meeting &x) {return (x.getSponsor() == userName || x.isParticipator(userName)) && x.getStartDate() < ed && x.getEndDate() > sd;}).empty())
-        return false;  // sponsor is busy
+    if (!m_storage->queryMeeting([userName, sd, ed](const Meeting &x) -> bool {
+        return ((x.getSponsor() == userName || x.isParticipator(userName))) && x.getStartDate() < ed && x.getEndDate() > sd;
+    }).empty()) return false;  // sponsor is busy
     for (auto n: participator) {
-        if (!m_storage->queryMeeting([n, sd, ed](const Meeting &x) {return (x.getSponsor() == n || x.isParticipator(n)) && x.getStartDate() < ed && x.getEndDate() > sd;}).empty())
-            return false;  // participator is busy
+        if (!m_storage->queryMeeting([n, sd, ed](const Meeting &x) -> bool {
+            return (x.getSponsor() == n || x.isParticipator(n)) && x.getStartDate() < ed && x.getEndDate() > sd;
+        }).empty()) return false;  // participator is busy
     }
     
     // all done
@@ -73,30 +92,62 @@ bool AgendaService::createMeeting(const std::string &userName, const std::string
     return true;
 }
 
+// ???
 bool AgendaService::addMeetingParticipator(const std::string &userName, const std::string &title, const std::string &participator) {
     if (userName == participator)
         return false;  // sponsor is participator
-    auto ls = m_storage->queryMeeting([userName, title](const Meeting &x) {return (x.getSponsor() == userName && x.getTitle() == title);});
+    
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return false;  // the sponsor does not exist
+    
+    if (m_storage->queryUser([participator](const User &x) -> bool {
+        return x.getName() == participator;
+    }).empty()) return false;  // the participator does not exist
+    
+    auto ls = m_storage->queryMeeting([userName, title](const Meeting &x) -> bool {
+        return (x.getSponsor() == userName && x.getTitle() == title);
+    });
     if (ls.size() != 1)
         return false;  // specific meeting not exist or duplicate meetings
     auto m = *(ls.begin());
     if (m.isParticipator(participator))
         return false;  // already in
-
-    if (!this->m_storage.get()->queryMeeting([participator, m](const Meeting &x) {return x.getSponsor() == participator && x.getStartDate() < m.getEndDate() && x.getEndDate() > m.getStartDate();}).empty())
-        return false;  // the participator sponsors a meeting that has conflict time
-    if (!this->m_storage.get()->queryMeeting([participator, m](const Meeting &x) {return x.isParticipator(participator) && x.getStartDate() < m.getEndDate() && x.getEndDate() > m.getStartDate();}).empty())
-        return false;  // the participator participates a meeting that has conflict time
-        
+    
+    if (m.isParticipator(userName))
+        return false;  // username is participator
+    
+    if (!m_storage->queryMeeting([participator, m](const Meeting &x) -> bool {
+        return x.getSponsor() == participator && x.getStartDate() < m.getEndDate() && x.getEndDate() > m.getStartDate();
+    }).empty()) return false;  // the participator sponsors a meeting that has conflict time
+    if (!m_storage->queryMeeting([participator, m](const Meeting &x) -> bool {
+        return x.isParticipator(participator) && x.getStartDate() < m.getEndDate() && x.getEndDate() > m.getStartDate();
+    }).empty()) return false;  // the participator participates a meeting that has conflict time
+    
     // all done
-    m_storage->updateMeeting([m](const Meeting &x) {return x.getTitle() == m.getTitle();}, [participator](Meeting &x) {x.addParticipator(participator);});
-    return true;
+    int updated = m_storage->updateMeeting([title, userName](const Meeting &x) -> bool {
+        return x.getSponsor() == userName && x.getTitle() == title;
+    }, [participator](Meeting &x) -> void {
+        x.addParticipator(participator);
+    });
+    return updated > 0;
 }
 
 bool AgendaService::removeMeetingParticipator(const std::string &userName, const std::string &title, const std::string &participator) {
     if (userName == participator)
         return false;  // sponsor cannot remove itself
-    auto ls = m_storage->queryMeeting([userName, title](const Meeting &x) {return x.getSponsor() == userName && x.getTitle() == title;});
+    
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return false;  // the sponsor does not exist
+    
+    if (m_storage->queryUser([participator](const User &x) -> bool {
+        return x.getName() == participator;
+    }).empty()) return false;  // the participator does not exist
+    
+    auto ls = m_storage->queryMeeting([userName, title](const Meeting &x) -> bool {
+        return x.getSponsor() == userName && x.getTitle() == title;
+    });
     if (ls.size() != 1)
         return false;  // specific meeting not exist or duplicate meetings
     auto m = *(ls.begin());
@@ -104,52 +155,121 @@ bool AgendaService::removeMeetingParticipator(const std::string &userName, const
         return false;  // not in the meeting
     
     // all done
-    m_storage->updateMeeting([title](const Meeting &x) {return x.getTitle() == title;}, [participator](Meeting &x) {x.removeParticipator(participator);});
-    return true;
+    int updated = m_storage->updateMeeting([userName, title](const Meeting &x) -> bool {
+        return x.getSponsor() == userName && x.getTitle() == title;
+    }, [participator](Meeting &x) -> void {
+        x.removeParticipator(participator);
+    });
+    
+    m_storage->deleteMeeting([](const Meeting &x) -> bool {
+        return x.getParticipator().empty();
+    });
+    return updated > 0;
 }
 
 bool AgendaService::quitMeeting(const std::string &userName, const std::string &title) {
-    if (m_storage->queryMeeting([userName, title](const Meeting &x) {return (x.getSponsor() != userName && x.isParticipator(userName) && x.getTitle() == title);}).size() != 1)
-        return false;  // not exist or duplicate
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return false;  // the sponsor does not exist
     
-    m_storage->updateMeeting([userName, title](const Meeting &x) {return (x.getSponsor() != userName && x.isParticipator(userName) && x.getTitle() == title);}, [userName](Meeting &x) {x.removeParticipator(userName);});
-    m_storage->deleteMeeting([](const Meeting &x) {return x.getParticipator().empty();});
-    return true;
+    if (m_storage->queryMeeting([userName, title](const Meeting &x) -> bool {
+        return (x.getSponsor() != userName && x.isParticipator(userName) && x.getTitle() == title);
+    }).empty()) return false;  // not exist or duplicate
+    
+    int updated = m_storage->updateMeeting([userName, title](const Meeting &x) -> bool {
+        return (x.getSponsor() != userName && x.isParticipator(userName) && x.getTitle() == title);
+    }, [userName](Meeting &x) -> void {
+        x.removeParticipator(userName);
+    });
+    
+    m_storage->deleteMeeting([](const Meeting &x) -> bool {
+        return x.getParticipator().empty();
+    });
+    return updated > 0;
 }
 
 std::list<Meeting> AgendaService::meetingQuery(const std::string &userName, const std::string &title) const {
-    return m_storage->queryMeeting([userName, title](const Meeting &x) {return x.getTitle() == title && (x.getSponsor() == userName || x.isParticipator(userName));});
+    std::list<Meeting> m;
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return m;  // the user does not exist, returning empty list
+    
+    m = m_storage->queryMeeting([userName, title](const Meeting &x) -> bool {
+        return x.getTitle() == title && (x.getSponsor() == userName || x.isParticipator(userName));
+    });
+    
+    return m;
 }
 
 std::list<Meeting> AgendaService::meetingQuery(const std::string &userName, const std::string &startDate, const std::string &endDate) const {
     Date sd(startDate), ed(endDate);
     std::list<Meeting> ml;
-    if (!Date::isValid(sd) || !Date::isValid(ed))
+    
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return ml;  // the user does not exist, returning empty list
+    
+    if (!Date::isValid(sd) || !Date::isValid(ed) || sd > ed)
         return ml;  // query date invalid
     
-    return m_storage->queryMeeting([userName, sd, ed](const Meeting &x) {return (x.getSponsor() == userName || x.isParticipator(userName)) && x.getEndDate() >= sd && x.getStartDate() <= ed;});
+    return m_storage->queryMeeting([userName, sd, ed](const Meeting &x) -> bool {
+        return (x.getSponsor() == userName || x.isParticipator(userName)) && x.getEndDate() >= sd && x.getStartDate() <= ed;
+    });
 }
 
 std::list<Meeting> AgendaService::listAllMeetings(const std::string &userName) const {
-    return this->m_storage->queryMeeting([userName](const Meeting &x) {return x.getSponsor() == userName || x.isParticipator(userName);});
+    std::list<Meeting> m;
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return m;  // the user does not exist, returning empty list
+    
+    return m_storage->queryMeeting([userName](const Meeting &x) -> bool {
+        return x.getSponsor() == userName || x.isParticipator(userName);
+    });
 }
 
 std::list<Meeting> AgendaService::listAllSponsorMeetings(const std::string &userName) const {
-    return m_storage->queryMeeting([userName](const Meeting &x) {return x.getSponsor() == userName;});
+    std::list<Meeting> m;
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return m;  // the user does not exist, returning empty list
+    
+    return m_storage->queryMeeting([userName](const Meeting &x) -> bool {
+        return x.getSponsor() == userName;
+    });
 }
 
 std::list<Meeting> AgendaService::listAllParticipateMeetings(const std::string &userName) const {
-    return m_storage->queryMeeting([userName](const Meeting &x) {return x.isParticipator(userName);});
+    std::list<Meeting> m;
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return m;  // the user does not exist, returning empty list
+    
+    return m_storage->queryMeeting([userName](const Meeting &x) -> bool {
+        return x.isParticipator(userName);
+    });
 }
 
 bool AgendaService::deleteMeeting(const std::string &userName, const std::string &title) {
-    m_storage->deleteMeeting([userName, title](const Meeting &x) {return (x.getSponsor() == userName) && (x.getTitle() == title);});
-    return true;
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return false;  // the user does not exist
+    
+    int deleted = m_storage->deleteMeeting([userName, title](const Meeting &x) -> bool {
+        return (x.getSponsor() == userName) && (x.getTitle() == title);
+    });
+    return deleted > 0;
 }
 
 bool AgendaService::deleteAllMeetings(const std::string &userName) {
-    m_storage->deleteMeeting([userName](const Meeting &x) {return x.getSponsor() == userName;});
-    return true;
+    if (m_storage->queryUser([userName](const User &x) -> bool {
+        return x.getName() == userName;
+    }).empty()) return false;  // the user does not exist
+    
+    int deleted = m_storage->deleteMeeting([userName](const Meeting &x) -> bool {
+        return x.getSponsor() == userName;
+    });
+    return deleted > 0;
 }
 
 void AgendaService::startAgenda(void) {
